pixelpwner/requirements.txt
astroid==2.3.3
autopep8==1.4.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.1
pycodestyle==2.5.0
pylint==2.4.4
six==1.13.0
typed-ast==1.4.0
wrapt==1.11.2
Image